+++ snowfall/src/main/java/com/jetradarmobile/snowfall/SnowfallView.kt
@@ -21,7 +21,6 @@ import android.graphics.Bitmap
 import android.graphics.Canvas
 import android.os.Handler
 import android.os.HandlerThread
-import android.support.v4.view.ViewCompat
 import android.util.AttributeSet
 import android.view.View
 
@@ -49,24 +48,26 @@ class SnowfallView(context: Context, attrs: AttributeSet) : View(context, attrs)
   private val snowflakesFadingEnabled: Boolean
   private val snowflakesAlreadyFalling: Boolean
 
-  private lateinit var snowflakes: Array<Snowflake>
   private val updateSnowflakesThread: UpdateSnowflakesThread
+  private var snowflakes: Array<Snowflake>? = null
 
   init {
     val a = context.obtainStyledAttributes(attrs, R.styleable.SnowfallView)
-    snowflakesNum = a.getInt(R.styleable.SnowfallView_snowflakesNum, DEFAULT_SNOWFLAKES_NUM)
-    snowflakeImage = a.getDrawable(R.styleable.SnowfallView_snowflakeImage)?.toBitmap()
-    snowflakeAlphaMin = a.getInt(R.styleable.SnowfallView_snowflakeAlphaMin, DEFAULT_SNOWFLAKE_ALPHA_MIN)
-    snowflakeAlphaMax = a.getInt(R.styleable.SnowfallView_snowflakeAlphaMax, DEFAULT_SNOWFLAKE_ALPHA_MAX)
-    snowflakeAngleMax = a.getInt(R.styleable.SnowfallView_snowflakeAngleMax, DEFAULT_SNOWFLAKE_ANGLE_MAX)
-    snowflakeSizeMinInPx = a.getDimensionPixelSize(R.styleable.SnowfallView_snowflakeSizeMin, dpToPx(DEFAULT_SNOWFLAKE_SIZE_MIN_IN_DP))
-    snowflakeSizeMaxInPx = a.getDimensionPixelSize(R.styleable.SnowfallView_snowflakeSizeMax, dpToPx(DEFAULT_SNOWFLAKE_SIZE_MAX_IN_DP))
-    snowflakeSpeedMin = a.getInt(R.styleable.SnowfallView_snowflakeSpeedMin, DEFAULT_SNOWFLAKE_SPEED_MIN)
-    snowflakeSpeedMax = a.getInt(R.styleable.SnowfallView_snowflakeSpeedMax, DEFAULT_SNOWFLAKE_SPEED_MAX)
-    snowflakesFadingEnabled = a.getBoolean(R.styleable.SnowfallView_snowflakesFadingEnabled, DEFAULT_SNOWFLAKES_FADING_ENABLED)
-    snowflakesAlreadyFalling = a.getBoolean(R.styleable.SnowfallView_snowflakesAlreadyFalling, DEFAULT_SNOWFLAKES_ALREADY_FALLING)
-    a.recycle()
-
+    try {
+      snowflakesNum = a.getInt(R.styleable.SnowfallView_snowflakesNum, DEFAULT_SNOWFLAKES_NUM)
+      snowflakeImage = a.getDrawable(R.styleable.SnowfallView_snowflakeImage)?.toBitmap()
+      snowflakeAlphaMin = a.getInt(R.styleable.SnowfallView_snowflakeAlphaMin, DEFAULT_SNOWFLAKE_ALPHA_MIN)
+      snowflakeAlphaMax = a.getInt(R.styleable.SnowfallView_snowflakeAlphaMax, DEFAULT_SNOWFLAKE_ALPHA_MAX)
+      snowflakeAngleMax = a.getInt(R.styleable.SnowfallView_snowflakeAngleMax, DEFAULT_SNOWFLAKE_ANGLE_MAX)
+      snowflakeSizeMinInPx = a.getDimensionPixelSize(R.styleable.SnowfallView_snowflakeSizeMin, dpToPx(DEFAULT_SNOWFLAKE_SIZE_MIN_IN_DP))
+      snowflakeSizeMaxInPx = a.getDimensionPixelSize(R.styleable.SnowfallView_snowflakeSizeMax, dpToPx(DEFAULT_SNOWFLAKE_SIZE_MAX_IN_DP))
+      snowflakeSpeedMin = a.getInt(R.styleable.SnowfallView_snowflakeSpeedMin, DEFAULT_SNOWFLAKE_SPEED_MIN)
+      snowflakeSpeedMax = a.getInt(R.styleable.SnowfallView_snowflakeSpeedMax, DEFAULT_SNOWFLAKE_SPEED_MAX)
+      snowflakesFadingEnabled = a.getBoolean(R.styleable.SnowfallView_snowflakesFadingEnabled, DEFAULT_SNOWFLAKES_FADING_ENABLED)
+      snowflakesAlreadyFalling = a.getBoolean(R.styleable.SnowfallView_snowflakesAlreadyFalling, DEFAULT_SNOWFLAKES_ALREADY_FALLING)
+    } finally {
+      a.recycle()
+    }
     updateSnowflakesThread = UpdateSnowflakesThread()
   }
 
@@ -76,29 +77,50 @@ class SnowfallView(context: Context, attrs: AttributeSet) : View(context, attrs)
 
   override fun onSizeChanged(w: Int, h: Int, oldw: Int, oldh: Int) {
     super.onSizeChanged(w, h, oldw, oldh)
-    val snowflakeParams = Snowflake.Params(
-        parentWidth = width, parentHeight = height, image = snowflakeImage,
-        alphaMin = snowflakeAlphaMin, alphaMax = snowflakeAlphaMax, angleMax = snowflakeAngleMax,
-        sizeMinInPx = snowflakeSizeMinInPx, sizeMaxInPx = snowflakeSizeMaxInPx,
-        speedMin = snowflakeSpeedMin, speedMax = snowflakeSpeedMax,
-        fadingEnabled = snowflakesFadingEnabled, alreadyFalling = snowflakesAlreadyFalling)
-    snowflakes = Array(snowflakesNum, { Snowflake(snowflakeParams) })
+    snowflakes = createSnowflakes()
+  }
+
+  override fun onVisibilityChanged(changedView: View, visibility: Int) {
+    super.onVisibilityChanged(changedView, visibility)
+    if (changedView === this && visibility == GONE) {
+      snowflakes?.forEach { it.reset() }
+    }
   }
 
   override fun onDraw(canvas: Canvas) {
     super.onDraw(canvas)
-    snowflakes.forEach { it.draw(canvas) }
+    if (isInEditMode) {
+      return
+    }
+    snowflakes?.forEach { it.draw(canvas) }
     updateSnowflakes()
+  }
+
+  private fun createSnowflakes(): Array<Snowflake> {
+    val snowflakeParams = Snowflake.Params(
+        parentWidth = width,
+        parentHeight = height,
+        image = snowflakeImage,
+        alphaMin = snowflakeAlphaMin,
+        alphaMax = snowflakeAlphaMax,
+        angleMax = snowflakeAngleMax,
+        sizeMinInPx = snowflakeSizeMinInPx,
+        sizeMaxInPx = snowflakeSizeMaxInPx,
+        speedMin = snowflakeSpeedMin,
+        speedMax = snowflakeSpeedMax,
+        fadingEnabled = snowflakesFadingEnabled,
+        alreadyFalling = snowflakesAlreadyFalling)
+    return Array(snowflakesNum, { Snowflake(snowflakeParams) })
   }
 
   private fun updateSnowflakes() {
     updateSnowflakesThread.handler.post {
-      snowflakes.forEach { it.update() }
-      ViewCompat.postInvalidateOnAnimation(this)
+      snowflakes?.forEach { it.update() }
+      postInvalidateOnAnimation()
     }
   }
 
-  private inner class UpdateSnowflakesThread() : HandlerThread("SnowflakesComputations") {
+  private inner class UpdateSnowflakesThread : HandlerThread("SnowflakesComputations") {
     val handler by lazy { Handler(looper) }
 
     init {